third_party/blink/web_tests/http/tests/websocket/send-file-blob-expected.txt

WebSocket: Send a File.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Got FileSystem object.
File created.
PASS fileEntry.isFile is true
Wrote to file.
Got File object.
PASS PASS: Message #0.
PASS closeEvent.wasClean is true
Deleting the file.
PASS successfullyParsed is true

TEST COMPLETE
